change name of the page files
