+++ src/pages/Login.js
@@ -1,0 +1,10 @@
+import Login from "../components/Authentication/Login";
+const LoginPage = () => {
+  return (
+    <section className="relative overflow-hidden">
+      <Login />
+    </section>
+  );
+};
+
+export default LoginPage;
