--- src/pages/LoginPage.js
@@ -1,10 +1,0 @@
-import Login from "../components/Authentication/Login";
-const LoginPage = () => {
-  return (
-    <section className="relative overflow-hidden">
-      <Login />
-    </section>
-  );
-};
-
-export default LoginPage;
