+++ src/pages/Signup.js
@@ -1,0 +1,10 @@
+import Signup from "../components/Authentication/Signup";
+const SignupPage = () => {
+  return (
+    <section className="relative overflow-hidden">
+      <Signup />
+    </section>
+  );
+};
+
+export default SignupPage;
