--- src/pages/SignupPage.js
@@ -1,10 +1,0 @@
-import Signup from "../components/Authentication/Signup";
-const SignupPage = () => {
-  return (
-    <section className="relative overflow-hidden">
-      <Signup />
-    </section>
-  );
-};
-
-export default SignupPage;
